+++ tests/conftest.py
@@ -10,7 +10,7 @@ os.environ["SENDER_EMAIL"] = "no-reply@test.com"
 os.environ["TO_EMAIL"] = "contact@test.com"
 os.environ["TO_NAME"] = "Test"
 os.environ["RECAPTCHA_ENABLED"] = "false"
-os.environ["MAILER_SERVICE"] = "sendgrid"
+os.environ["MAILER_PROVIDER"] = "sendgrid"
 os.environ["SENDGRID_API_KEY"] = "1234567890"
 os.environ["SENDGRID_SANDBOX"] = "true"